requirements.txt
absl-py==2.1.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.4
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.25.1
marshmallow-sqlalchemy==1.4.0
matplotlib==3.10.0
mediapipe==0.10.18
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.5
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
SQLAlchemy==2.0.37
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.66
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
